# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LanGigaCards API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Quiz/questions": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "lessonId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "count",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Quiz/sessions": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartQuizSessionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartQuizSessionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartQuizSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Quiz/sessions/{sessionId}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Quiz/sessions/{sessionId}/answers": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizAnswerDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizAnswerDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Quiz/card-sessions": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCardQuizDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCardQuizDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCardQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Quiz/card-completion": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "deckId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "languageCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardQuizAnswerDto": {
        "type": "object",
        "properties": {
          "wordId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "skipped": {
            "type": "boolean"
          },
          "timeSpentSeconds": {
            "maximum": 3600,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "StartQuizSessionDto": {
        "type": "object",
        "properties": {
          "lessonId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "questionCount": {
            "maximum": 50,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SubmitCardQuizDto": {
        "type": "object",
        "properties": {
          "deckId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "languageCode": {
            "maxLength": 10,
            "type": "string",
            "nullable": true
          },
          "answers": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardQuizAnswerDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SubmitQuizAnswerDto": {
        "type": "object",
        "properties": {
          "quizId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "selectedOptionId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "skip": {
            "type": "boolean"
          },
          "timeSpentSeconds": {
            "maximum": 3600,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Token değerinizi girin.",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}